processing_layer/ai_analysis/requirements.txt
fastapi
uvicorn
google-generativeai
kafka-python
